Clear/complete recent TODOs
I shortened `format_reply`. The `create` method is as simple as I can think to make it for now, and it works well enough. To create a model for `totals` would be involved, since I would ideally have to use a celery beat (and therefore LEARN celery) to have the totals object be created on schedule, and I'm not ready for that.